genai-agents-backend/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# LangChain and AI
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langchain-anthropic==0.2.4
langchain-openai==0.2.5
google-generativeai>=0.7.0,<0.8.0
anthropic==0.39.0
openai==1.54.4

# BigQuery and Google Cloud
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
sqlalchemy-bigquery==1.11.0
db-dtypes==1.3.0

# Authentication
pyjwt==2.9.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.0

# Async and streaming
asyncio==3.4.3
aiofiles==24.1.0
sse-starlette==2.1.0

# Monitoring and logging
python-json-logger==2.0.7
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-gcp-trace==1.9.0

# Data processing
pandas==2.2.2
numpy<2.0.0
scipy==1.13.1
python-dateutil==2.8.2

# Development
pytest==8.3.2
pytest-asyncio==0.24.0
httpx==0.27.0
black==24.8.0
pylint==3.2.6